Optimize pinned memory allocation in CudaBasic.
This PR introduces CudaHostAllocator, which allocates a single slab of
pinned host memory. It works by providing fixed-size chunks of memory
for use by the CudaBasic channel.

The CudaBasic channel itself is being reworked to chunk large input
tensors (so that each chunk fits into a pinned memory chunk), and send
each one separately, in order, over the the underlying CPU channel.

NOTE: At the moment, the CudaHostAllocator will block a whole chunk
for each allocation, possibly leading to under-used memory (for
instance when sending several small tensors). It is possible to
mitigate this by fitting multiple allocations per chunk by maintaining
two variables per chunk: `offset` and `used`, the former being
non-decreasing. Each time an allocation for `size` bytes is requested,
each chunk is checked, and if `chunkSize-offset >= size`, `offset` and
`used` are both incremented by `size`. When an allocation is being
reclaimed, the `used` variable is decremented by `size`, and if it
falls to zero, `offset` is reset to zero as well.

ghstack-source-id: 0296fc470c98a86324016365cc98669af0769484
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/303